requirements.txt
pyboy==2.6.0
PyOpenGL
pillow
opencv-python
numpy
langchain
langgraph
openai
ollama